Only set logging config on executable files
This mostly prevents the main module from logging because some random
component deep down decided to change the log level or some other
setting. In the future we need to use proper logger classes too, so we
can know where the logging data is coming from. It is an absolute mess
right now.