When agent gets tainted, give it a negative score
this mostly prevents tainted agents to draw against agents with score 0.

With a negative score only being attainable when flagged as tainted,
this ensure that a tainted agent wll never draw with a non-tainted one.
Two tainted agents can still draw, if it happens at the same time,
otherwise the one that gets tainted first loses.